tg_api_old/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.121.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyngrok==7.4.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.49.3
Telethon==1.41.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
